+++ priv/repo/migrations/20220505081053_change_reference_parent_id_rule_results.exs
@@ -1,0 +1,11 @@
+defmodule TdDd.Repo.Migrations.ChangeReferenceParentIdRuleResults do
+  use Ecto.Migration
+
+  def change do
+    alter table("rule_results") do
+      modify(:parent_id, references("rule_results", on_delete: :delete_all),
+        from: references("rule_results", on_delete: :delete_all)
+      )
+    end
+  end
+end