+++ lib/td_dq_web/controllers/implementation_controller.ex
@@ -148,7 +148,8 @@ defmodule TdDqWeb.ImplementationController do
       |> filter_links_by_permission(claims)
       |> filter_data_structures_by_permission(claims)
 
-    actions = build_implementation_actions(claims, implementation)
+      actions = claims
+      |> build_implementation_actions(implementation)
       |> put_edit_action(claims, implementation)
 
     with {:can, true} <- {:can, can?(claims, show(implementation))} do
@@ -173,6 +174,7 @@ defmodule TdDqWeb.ImplementationController do
 
   defp build_implementation_actions(claims, implementation) do
     [
+      :execute,
       :link_concept,
       :link_structure,
       :manage,