+++ CHANGELOG.md
@@ -6,6 +6,10 @@
 
 - [TD-4045] Rule implementation `execute` action
 
+### Fixed
+
+- [TD-4799] Performance issue filtering executions by source
+
 ## [4.43.1] 2022-05-11
 
 - [TD-4089] Support for Implementations without Rules